Revert to having `State.sample()` as the abstractmethod
